Refactor: Extract `renderScreenshot` into Screenshot.js
